machine_learning/classification/logistic_regression/char_reco/requirements.txt
streamlit
numpy
scikit-learn
opencv-python-headless
Pillow
streamlit-drawable-canvas
